# app/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "InTy",
        "description": "InTy",
        "version": "1.0.2"
    },
    "paths": {
...
        "/api/v1/users/profile": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get Profile",
                "description": "Get current user profile.",
                "operationId": "get_profile_api_v1_users_profile_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Update Profile",
                "description": "Update current user profile.",
                "operationId": "update_profile_api_v1_users_profile_put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/users/avatar": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Upload Avatar",
                "operationId": "upload_avatar_api_v1_users_avatar_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_avatar_api_v1_users_avatar_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse_User_"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/users/device/register": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Register Device Token",
                "description": "注册或更新设备token",
                "operationId": "register_device_token_api_v1_users_device_register_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeviceTokenRegister"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/users/deletion/check": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Check Deletion Eligibility",
                "description": "检查用户是否可以删除账户",
                "operationId": "check_deletion_eligibility_api_v1_users_deletion_check_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse_DeletionCheckResponse_"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/users/delete-account": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Delete User Account",
                "description": "删除用户账户",
                "operationId": "delete_user_account_api_v1_users_delete_account_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "anyOf": [
                                    {
                                        "$ref": "#/components/schemas/AccountDeletionRequest"
                                    },
                                    {
                                        "type": "null"
                                    }
                                ],
                                "title": "Request"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse_AccountDeletionResponse_"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "APIResponse": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "success"
                    },
                    "data": {
                        "anyOf": [
                            {},
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Data"
                    }
                },
                "type": "object",
                "title": "APIResponse"
            },
            "APIResponse_AccountDeletionResponse_": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "success"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/AccountDeletionResponse"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "title": "APIResponse[AccountDeletionResponse]"
            },
...
            "APIResponse_DeletionCheckResponse_": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "success"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/DeletionCheckResponse"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "title": "APIResponse[DeletionCheckResponse]"
            },
...
            "APIResponse_User_": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "success"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "title": "APIResponse[User]"
            },
...
            "AccountDeletionRequest": {
                "properties": {
                    "reason": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 255
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Reason",
                        "description": "删除原因"
                    }
                },
                "type": "object",
                "title": "AccountDeletionRequest",
                "description": "账户删除请求",
                "example": {
                    "reason": "隐私关注"
                }
            },
...
            "Body_upload_avatar_api_v1_users_avatar_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_upload_avatar_api_v1_users_avatar_post"
            },
...
            "DeviceTokenRegister": {
                "properties": {
                    "token": {
                        "type": "string",
                        "title": "Token"
                    }
                },
                "type": "object",
                "required": [
                    "token"
                ],
                "title": "DeviceTokenRegister",
                "description": "设备token注册请求"
            },
...
            "Gender": {
                "type": "string",
                "enum": [
                    "MALE",
                    "FEMALE",
                    "OTHER"
                ],
                "title": "Gender",
                "description": "性别"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "User": {
                "properties": {
                    "readable_id": {
                        "type": "string",
                        "title": "Readable Id"
                    },
                    "nickname": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Nickname"
                    },
                    "avatar": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Avatar"
                    },
                    "email": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Email"
                    },
                    "phone": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Phone"
                    },
                    "gender": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Gender"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "age_group": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Age Group"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "system_language": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "System Language"
                    },
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "auth_type": {
                        "type": "string",
                        "title": "Auth Type"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Updated At"
                    },
                    "is_superuser": {
                        "type": "boolean",
                        "title": "Is Superuser",
                        "default": false
                    },
                    "public_agents_count": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Public Agents Count",
                        "default": 0
                    },
                    "total_public_agents_follows": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Total Public Agents Follows",
                        "default": 0
                    },
                    "followers_count": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Followers Count",
                        "default": 0
                    },
                    "connector_count": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Connector Count",
                        "default": 0
                    }
                },
                "type": "object",
                "required": [
                    "readable_id",
                    "id",
                    "auth_type",
                    "is_active",
                    "created_at"
                ],
                "title": "User",
                "description": "返回给客户端的用户信息"
            },
...
            "UserUpdate": {
                "properties": {
                    "nickname": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Nickname"
                    },
                    "avatar": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Avatar"
                    },
                    "email": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Email"
                    },
                    "phone": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Phone"
                    },
                    "gender": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Gender"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "age_group": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Age Group"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "system_language": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "System Language"
                    }
                },
                "type": "object",
                "title": "UserUpdate",
                "description": "更新用户信息"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "\n            输入格式为: your_token\n            注意: 不需要Bearer前缀\n            "
            }
        }
    }
}